// src/renderer/src/types/dental-api.d.ts
export interface Patient {
  id: number;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  gender: string;
  phone: string;
  email: string | null;
  address: string | null;
  insuranceProvider: string | null;
  insuranceNumber: string | null;
  medicalHistory: string | null;
  allergies: string | null;
  notes: string | null;
  createdAt: number | Date;
  updatedAt: number | Date;
}